refactor(phase-d): consolidate Coloc — features/coloc/ + mixin-composed extension package
Align coloc with the vlrgg/zevent/minecraft pattern:

- src/coloc/ -> features/coloc/ (pure logic: api_client, constants, models,
  storage, utils). Relative internal imports are unchanged.
- extensions/coloc.py -> extensions/coloc/ package split by concern:
  * _common.py    — ColocConfig schema (still registered as `moduleColoc` for
                    data compat), logger, module-level config
  * reminders.py  — RemindersMixin: daily journa check, /journa set/remove,
                    reminder_checker, advent calendar nudges
  * events.py     — EventsMixin: hourly Zunivers event poll, hardcore season
                    transitions, /zunivers check
  * corporation.py — CorporationMixin: daily recap + /corpo command
  * __init__.py   — ColocExtension composition, fun commands (/fesse,
                    /massageducul), on_startup wiring, setup(bot)

All cross-file imports switched from src.coloc.* to features.coloc.* (the
feature package's __init__ re-exports the public API, so most call sites use
`from features.coloc import X`).

The `moduleColoc` config key is intentionally preserved to avoid a MongoDB
migration; a follow-up PR can split the Zunivers logic into its own extension
and rename the module once migration is in scope.
